A bunch of changes related to level 2 development
- Interaction/hint papers:
Removed colored fonts in favor of a color setting in the component
Made the paper more natural how you'd hold it irl
Made the papers use the menu system so you don't need to reimplement things it already does

- Player interaction has been overhauled and no longer uses the collision triggers instead using a raycast from the player. This was done due to wanting to add more hidden hints and not need to code enabling/disabling but have a wall block the raycast instead. Added interaction text and distance based scaling.

- Removed a few unused scripts

- Removed passwords from activatable and added the new lock system which allows defining locks that aren't just passwords.

+++ Project/Assets/CET/Scripts/VPL/ValueConverters/PowConverter.cs
@@ -8,7 +8,7 @@ public class PowConverter : ValueConverter
     {
         if (input[0] is float lfloat && input[1] is float rfloat)
         {
-            return Math.Pow(lfloat, rfloat);
+            return (float)Math.Pow(lfloat, rfloat);
         }
 
         throw new Exception("Power is only defined for numbers!");
